+++ sql/updates/Object_20131001_1.sql
@@ -1,6 +1,6 @@
 
 CREATE TABLE IF NOT EXISTS `team_pending_player` (
-  `teamId` int(10) unsigned NOT NULL,
+  `teamId` bigint(20) unsigned NOT NULL,
   `playerId` bigint(20) unsigned NOT NULL,
   `opTime` int(10) unsigned NOT NULL,
   PRIMARY KEY (`teamId`,`playerId`)
